requirements.txt
Django==5.1
torch==2.5.1+cpu
torchvision==0.20.1+cpu
torchaudio==2.5.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
transformers==4.44.2
Pillow==10.0.0
gTTS==2.3.2
googletrans==4.0.0-rc1
sentencepiece==0.2.1
accelerate
gunicorn==22.0.0
numpy==2.1.3
scikit-learn==1.5.2
mysqlclient==2.2.4
